+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * [BC BREAK] Rename `SourceMap` to `SourceCollection`, its methods from `getSources()` to `all()` and `addSource()` to `add()`
  * [BC BREAK] Third Argument of `ToolResult::__construct()` now expects `SourceCollection` instead of `array<int, Source>`
  * Add `maxToolCalls` parameter to `AgentProcessor` to limit tool calling iterations and prevent infinite loops
+ * Add `Countable` and `IteratorAggregate` implementations to `SourceCollection`
 
 0.2
 ---